:art: re-organizes the analyticsTracking service
adds an index.js to allow for multiple tracking providers to be implemented from within the analyticsTracking service
